chore: ディレクトリ名を複数形に変更
